Cleaned up ch.utils.
