+++ cherry/models/__init__.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+from .utils import polyak_average
 from . import atari
 from . import control
 from . import tabular
